UNTESTED: Add Frequency locked loop block
Add frequency locked loop to compensate doppler shift based on
averaged inter-sample phase difference. The frequency estimate
is averaged over the whole block and all samples are frequency
compensated on the basis of this estimate independent of the
block size. If blocks are aligned volk is used where possible.

Code is compiling but untested. (End of day commit)