Remove double/float switch from lpf2, PLL compensates rotated QPSK constellation
lpf2 uses doubles instead of defined constants to switch between float
and double as double is still fast enough and much easier to work with.

PLL rotates the M-Power samples to the real axis. QPSK constellation is
shifted by 45 degree first, to put the zero phase constellation to the
real axis as well.