Add M-PSK Ambiguity Resolver, Share common code between M-PSK Modulator and the two Demodulators, Improve modulation/demodulation testcases
Add M-PSK Ambiguity Resolver based on the attached sync marker

Share common code between M-PSK Modulator and the two Demodulators
Each block had an own copy of the Constellation generator. Now all
three use the same code from a shared header file.

Improve modulation/demodulation testcases
Beside pregenerated testcases, each test will take the output or the
Modulator and demodulate it and compare it to the input to make sure
that the input to the Modulator is the same as the output of the
Demodulator with a perfect channel.

Fix some unused header includes.