Remove legacy SWIG code
Remove swig code for the convolutional_encode helper class, that was
once meant to be a separate block, but never was. It is only used by
some classes internaly but never needs to be exported to python.
Some SWIG leftovers caused the compilation to fail on modern systems,
thus removing them.

The Convolutional encoder block of the gr-ccsds package is not
affected at all by this change as it does not use this helper class.